api: rename types->trezortypes
