+++ internal/core/core.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/trezor/trezord-go/internal/logs"
 	"github.com/trezor/trezord-go/internal/message"
-	"github.com/trezor/trezord-go/types"
+	types "github.com/trezor/trezord-go/trezorapi/trezortypes"
 )
 
 // Package with "core logic" of device listing
