+++ internal/server/status/status.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/trezor/trezord-go/internal/logs"
 	"github.com/trezor/trezord-go/trezorapi"
-	"github.com/trezor/trezord-go/types"
+	types "github.com/trezor/trezord-go/trezorapi/trezortypes"
 
 	"github.com/gorilla/csrf"
 	"github.com/gorilla/mux"
