+++ trezorapi/trezortypes/types.go
@@ -1,4 +1,4 @@
-package types
+package trezortypes
 
 type DeviceType int
